complited CRUD posts

+++ app/Models/Post.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use Carbon\Carbon;
 use Cviebrock\EloquentSluggable\Sluggable;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
@@ -53,15 +54,15 @@ class Post extends Model
     const IS_PUBLIC = 1;
     const IS_FEATURED = 1;
     const IS_STANDART = 0;
-    protected $fillable = ['title', 'content'];
+    protected $fillable = ['title', 'content', 'date'];
 
     public function category()
     {
-        return $this->hasOne(Category::class);
+        return $this->belongsTo(Category::class);
     }
     public function author()
     {
-        return $this->hasOne(User::class);
+        return $this->belongsTo(User::class, 'user_id');
     }
     public function tags()
     {
@@ -100,17 +101,24 @@ class Post extends Model
 
     public function remove()
     {
-        Storage::delete('uploads/' . $this->image);
+        $this->removeImage();
         $this->delete();
     }
 
+    public function removeImage()
+    {
+        if($this->image != null){
+
+            Storage::delete('uploads/' . $this->image);
+        }
+    }
     public function uploadImage($image)
     {
         if($image == null) { return; }
 
-        Storage::delete('uploads/' . $this->image);
-        $filename = Str::random(10) . '.' . $image->extention();
-        $image->saveAs('uploads', $filename);
+        $this->removeImage();
+        $filename = Str::random(10) . '.' . $image->extension();
+        $image->storeAs('uploads', $filename);
         $this->image = $filename;
         $this->save();
     }
@@ -119,7 +127,7 @@ class Post extends Model
     {
         if($this->image == null)
         {
-            return 'img\no-image.png';
+            return '/img/no-image.png';
         }
         return '/uploads/' . $this->image;
     }
@@ -171,6 +179,12 @@ class Post extends Model
         $this->save();
     }
 
+    public function setFeatured()
+    {
+        $this->is_featured = Post::IS_FEATURED;
+        $this->save();
+    }
+
     public function toggleFeature($value)
     {
         if($value == null)
@@ -178,6 +192,31 @@ class Post extends Model
             return $this->setStandart();
         }
         return $this->setFeatured();
+    }
+
+    public function setDateAttribute($value)
+    {
+        $date = Carbon::createFromFormat('d/m/y', $value)->format('Y-m-d');
+        $this->attributes['date'] = $date;
+    }
+
+    public function getDateAttribute($value)
+    {
+        $date = Carbon::createFromFormat('Y-m-d', $value)->format('d/m/y');
+        return $date;
+    }
+
+    public function getCategoryTitle()
+    {
+        return ($this->category != null)
+            ? $this->category->title
+            : 'Нет категории';
+    }
+    public function getTagsTitles()
+    {
+        return (!$this->tags->isEmpty())
+            ? implode(', ', $this->tags->pluck('title')->all())
+            : 'Нет тегов';
     }
 }
